requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2==2.8.6
requests==2.31.0
bs4==0.0.1
celery[redis]==5.2.7
flower==1.2.0
lxml==4.9.3
django-celery-beat
djangorestframework-simplejwt


